requirements.txt
alembic==1.8.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
databases==0.6.2
ecdsa==0.18.0
fastapi==0.87.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
Naked==0.1.32
passlib==1.7.4
Pillow==9.3.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.4.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.21.0
status==0.2.5
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
